Moved malaria functionality to its own separate Django app
