+++ infohub/settings.py
@@ -29,6 +29,7 @@ INSTALLED_APPS = (
     'rest_framework',
     'rest_framework_swagger',
     'infohub',
+    'malaria',
     'peacetrack',
     'webhub',
 )
