+++ infohub/urls.py
@@ -7,6 +7,7 @@ urlpatterns = patterns(
     url(r'^admin/', include(admin.site.urls)),
     url(r'^', include('webhub.urls')),
     url(r'^docs/', include('rest_framework_swagger.urls')),
+    url(r'^malaria/', include('malaria.urls', namespace='malaria')),
     url(r'^peacetrack/', include('peacetrack.urls')),
     url(r'^webhub/', include('webhub.urls', namespace='webhub')),
 )
