+++ malaria/templates/malaria/notice.html
@@ -7,7 +7,7 @@
     <body>
         <br><br><br><br><br><br><br><br>
         <center><p><h4> {{ text }} </h4></p></center>
-        <center><p><h4><a href="{% url 'webhub:view_post' post.id %}"><u> {{ text1 }} </u></a></h4></p></center>
+        <center><p><h4><a href="{% url 'malaria:view_post' post.id %}"><u> {{ text1 }} </u></a></h4></p></center>
         <br><br><br><br><br><br><br><br><br><br><br><br><br>
     </body>
 {% endblock %}
