+++ malaria/templates/malaria/view_post.html
@@ -33,8 +33,8 @@
         </div><br>
         <br>    
         <center>
-            <a href="{% url 'webhub:edit_post' post.id %}" class="btn btn-default">Edit</a>
-            <a href="{% url 'webhub:delete_post' post.id %}" class="btn btn-default">Delete</a>
+            <a href="{% url 'malaria:edit_post' post.id %}" class="btn btn-default">Edit</a>
+            <a href="{% url 'malaria:delete_post' post.id %}" class="btn btn-default">Delete</a>
         </center>
         <br>
     </div>
