+++ malaria/tests.py
@@ -6,7 +6,8 @@ from rest_framework import status
 from rest_framework.parsers import JSONParser
 from rest_framework.renderers import JSONRenderer
 from rest_framework.test import APITestCase
-from webhub.models import Pcuser, Post
+from malaria.models import Post
+from webhub.models import Pcuser
 from webhub.serializers import PostSerializer
