+++ peacetrack/urls.py
@@ -10,7 +10,6 @@ from peacetrack import views
 
 urlpatterns = patterns('',
     url(r'^$', views.peacetrack, name='peacetrack'),
-    url(r'^malaria/$', webhub_views.malaria, name='malaria'),
 #    url(r'^api-auth/', include('rest_framework.urls', namespace='rest_framework')),
 #    url(r'^api/', include(router.urls)),
     url(r'^volunteer/$', views.volunteer, name='volunteer'),
