+++ ui/dashboard.html
@@ -24,14 +24,14 @@ email : ranihaileydesai@gmail.com
     <div class="col-md-6">
     <div style="width:600px; height:300px; margin:0 auto; background-color:#0d8b97;">
     <br>
-        <center><h2><a href="/malaria/"><font color="white"><u>Malaria</u></a></h2></center>
+        <center><h2><a href="/malaria/list_posts/"><font color="white"><u>Malaria</u></a></h2></center>
     
     <div class="row">
         <div class="col-md-1">
         </div>
         <div class="col-md-5">
             <br>Tiborem udělá nám Helenka nakrájí a štyry pláty, pěkně nahrubo, Láďa se vaří. Trochu mouky jen zezhora to spojit. Nudle sou dozlatova.</font><br><br>
-            <button class="btn" onclick="window.location.href='/malaria/'">Go to Malaria</button>
+            <button class="btn" onclick="window.location.href='/malaria/list_posts/'">Go to Malaria</button>
         </div>
         <div class="col-md-6">
             <img src="http://i.imgur.com/y8lfLmz.jpg" class="img-responsive" style="width:240px; height:200px; padding-bottom:5px; padding-top:10px;">
