+++ webhub/serializers.py
@@ -1,5 +1,6 @@
 from django.contrib.auth.models import User
 from webhub.models import *
+from malaria.models import Post, RevPost
 from peacetrack.models import *
 
 from rest_framework import serializers
